pyproject.toml
[project]
name = "mss-metro-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.2.0",
    "aiohttp>=3.10.0",
    "fastapi>=0.118.0",
    "mangum>=0.19.0",
    "pillow>=11.0.0",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "python-ulid>=3.0.0",
    "structlog>=25.4.0",
    # Redis for caching
    "redis>=5.0.0",
    # Celery for async task processing
    "celery[redis]>=5.3.0",
    # OpenSearch for vector storage
    "opensearch-py>=2.4.0",
    # LangChain for structured queries
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-core>=0.1.0",
    # VLM and ML dependencies
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.16.0",
    "optimum>=1.16.0",
    "sentence-transformers>=2.2.0",
    # Additional ML utilities
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "scikit-image>=0.22.0",
    "dependency-injector>=4.42.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.13.3",
    "taskipy>=1.14.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.pyc",
    ".pytest_cache",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RET", # flake8-return
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function calls in argument defaults
    "RET504", # unnecessary variable assignment before return
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
explicit_package_bases = true
namespace_packages = true

[[tool.mypy.overrides]]
module = [
    "aioboto3.*",
    "pymemcache.*",
    "mangum.*",
    "ulid.*",
]
ignore_missing_imports = true

[tool.taskipy.tasks]
lint = "ruff check app/ tests/"
lint-fix = "ruff check --fix app/ tests/ && ruff format app/ tests/"
format = "ruff format app/ tests/"
type-check = "mypy app/"
test = "pytest tests/ -v"
test-cov = "pytest tests/ --cov=app --cov-report=term-missing --cov-report=html"
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
pre-commit = "pre-commit run --all-files"
pre-commit-install = "pre-commit install"
pre-commit-update = "pre-commit autoupdate"
ci = "task lint && task type-check && task test"
clean = "find . -type d -name '__pycache__' -exec rm -rf {} + 2>/dev/null || true && find . -type f -name '*.pyc' -delete && rm -rf .pytest_cache .ruff_cache .mypy_cache htmlcov .coverage"
